feat(webui): ensure assets are fetched from the OVPM HTTP server only
Previously we were fetching the necessary webui libraries such as bootstrap
and mui css, from the CDN for the sake of bandwidth and round-trip time.

But now we have decided to serve these assets from the OVPM server itself for
better integrity and protection against
information leaks such as IP address of the user agent to the 3rd party CDN
services.

Closes #65